1. Fix: now activity text will have the integration it is created from 2. Fix: better logging in beautify output
