+++ integrations/github/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sigma/github",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "description": "github extension for Sigma",
   "files": [
     "github"
